+++ tests/core/tests/widgets_tests.py
@@ -1,16 +1,17 @@
+# -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
 from decimal import Decimal
-from datetime import date
+from datetime import date, datetime
 
 from django.test import TestCase
 
 from import_export import widgets
 
 from core.models import (
-        Author,
-        Category,
-        )
+    Author,
+    Category,
+)
 
 
 class BooleanWidgetTest(TestCase):
@@ -21,6 +22,10 @@ class BooleanWidgetTest(TestCase):
     def test_clean(self):
         self.assertTrue(self.widget.clean("1"))
         self.assertTrue(self.widget.clean(1))
+        self.assertEqual(self.widget.clean(""), None)
+
+    def test_render(self):
+        self.assertEqual(self.widget.render(None), "")
 
 
 class DateWidgetTest(TestCase):
@@ -32,15 +37,55 @@ class DateWidgetTest(TestCase):
     def test_render(self):
         self.assertEqual(self.widget.render(self.date), "13.08.2012")
 
+    def test_render_none(self):
+        self.assertEqual(self.widget.render(None), "")
+
     def test_clean(self):
         self.assertEqual(self.widget.clean("13.08.2012"), self.date)
 
 
-class DecimalWidgetTest(TestCase):
+class DateTimeWidgetTest(TestCase):
+
+    def setUp(self):
+        self.datetime = datetime(2012, 8, 13, 18, 0, 0)
+        self.widget = widgets.DateTimeWidget('%d.%m.%Y %H:%M:%S')
+
+    def test_render(self):
+        self.assertEqual(self.widget.render(self.datetime),
+                         "13.08.2012 18:00:00")
+
+    def test_render_none(self):
+        self.assertEqual(self.widget.render(None), "")
 
     def test_clean(self):
-        widget = widgets.DecimalWidget()
-        self.assertEqual(widget.clean("11.111"), Decimal("11.111"))
+        self.assertEqual(self.widget.clean("13.08.2012 18:00:00"),
+                         self.datetime)
+
+
+class DateWidgetBefore1900Test(TestCase):
+
+    def setUp(self):
+        self.date = date(1868, 8, 13)
+        self.widget = widgets.DateWidget('%d.%m.%Y')
+
+    def test_render(self):
+        self.assertEqual(self.widget.render(self.date), "13.08.1868")
+
+    def test_clean(self):
+        self.assertEqual(self.widget.clean("13.08.1868"), self.date)
+
+
+class DecimalWidgetTest(TestCase):
+
+    def setUp(self):
+        self.value = Decimal("11.111")
+        self.widget = widgets.DecimalWidget()
+
+    def test_clean(self):
+        self.assertEqual(self.widget.clean("11.111"), self.value)
+
+    def test_render(self):
+        self.assertEqual(self.widget.render(self.value), self.value)
 
 
 class ForeignKeyWidgetTest(TestCase):
@@ -66,7 +111,8 @@ class ManyToManyWidget(TestCase):
 
     def setUp(self):
         self.widget = widgets.ManyToManyWidget(Category)
-        self.cat1 = Category.objects.create(name='Cat 1')
+        self.widget_name = widgets.ManyToManyWidget(Category, field="name")
+        self.cat1 = Category.objects.create(name=u'Cat úňíčóďě')
         self.cat2 = Category.objects.create(name='Cat 2')
 
     def test_clean(self):
@@ -79,3 +125,6 @@ class ManyToManyWidget(TestCase):
     def test_render(self):
         self.assertEqual(self.widget.render(Category.objects),
                 "%s,%s" % (self.cat1.pk, self.cat2.pk))
+        self.assertEqual(self.widget_name.render(Category.objects),
+                u"%s,%s" % (self.cat1.name, self.cat2.name))
+