+++ docs/changelog.rst
@@ -1,10 +1,65 @@
 Changelog for django-import-export
 ==================================
 
-0.2.2 (unreleased)
+0.2.7 (unreleased)
 ------------------
 
 - Nothing changed yet.
+
+
+0.2.6 (2014-10-09)
+------------------
+
+- added use of get_diff_headers method into import.html template (#158)
+
+- Try to use OrderedDict instead of SortedDict, which is deprecated in 1
+  django 1.7 (#157)
+
+- fixed #105 unicode import
+
+- remove invalid form action "form_url" #154
+
+
+0.2.5 (2014-10-04)
+------------------
+
+- Do not convert numeric types to string (#149)
+
+- implement export as an admin action (#124)
+
+
+0.2.4 (2014-09-18)
+------------------
+
+- fix: get_value raised attribute error on model method call
+
+- Fixed XLS import on python 3. Optimized loop
+
+- Fixed properly skipping row marked as skipped when importing data from 
+  the admin interface.
+  
+- Allow Resource.export to accept iterables as well as querysets
+
+- Improve error messages
+
+- FIX: Properly handle NullBoleanField (#115) - Backward Incompatible Change
+  previously None values were handled as false
+
+
+0.2.3 (2014-07-01)
+------------------
+
+- Add separator and field keyword arguments to ManyToManyWidget
+
+- FIX: No support for dates before 1900 (#93)
+
+
+0.2.2 (2014-04-18)
+------------------
+
+- RowResult now stores exception object rather than it's repr
+
+- Admin integration - add EntryLog object for each added/updated/deleted instance
 
 
 0.2.1 (2014-02-20)